requirements.txt
anyio==4.3.0
astor==0.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
duckdb==0.9.2
executing==2.0.1
Faker==19.13.0
fonttools==4.49.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
ipython==8.21.0
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==1.5.3
pandasai==1.5.19
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.14
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.1
wcwidth==0.2.13
